First Version of Computing, not tested yet

+++ src/main/kotlin/com/triangl/trackingIngestion/entity/InputDataPoint.kt
@@ -17,7 +17,7 @@ class InputDataPoint {
     var deviceId: String? = null
 
     @NotNull
-    var signalStrength: Float? = null
+    var signalStrength: Int? = null
 
     @NotNull
     var timeOfFlight: Long? = null
